move apiHooks into its own top level folder
