+++ src/redux-ducks/library/index.js
@@ -13,7 +13,7 @@ import {
 } from "redux-ducks/chapters/actions";
 import { selectSources } from "redux-ducks/sources";
 import { selectCategoryMangaIds } from "redux-ducks/categories";
-import { type UnreadMap } from "components/apiHooks";
+import { type UnreadMap } from "apiHooks";
 import filterSortLibrary from "./libraryUtils";
 import {
   FETCH_LIBRARY,
